mammo_classification/requirements.txt
# Deep Learning
torch>=1.9.0
torchvision>=0.10.0
timm>=0.6.12

# Image Processing
Pillow>=8.3.1
albumentations>=1.4.23

# Data Science & Metrics
numpy>=1.21.2
pandas>=1.3.3
scikit-learn>=0.24.2

# Configuration
PyYAML>=5.4.1

# Visualization
matplotlib>=3.4.3
seaborn>=0.11.2

# Utilities
tqdm>=4.62.3